+++ samples/chart.go
@@ -21,7 +21,7 @@ func ChartBar() {
 		},
 	}
 
-	gouse.CreateBarChart(newChart)
+	gouse.BarChart(newChart)
 }
 
 /*
@@ -40,7 +40,7 @@ func ChartLine() {
 		},
 	}
 
-	gouse.CreateLineChart(newChart)
+	gouse.LineChart(newChart)
 }
 
 /*
@@ -65,5 +65,5 @@ func ChartPie() {
 		},
 	}
 
-	gouse.CreatePieChart(newChart)
+	gouse.PieChart(newChart)
 }