+++ samples/net.go
@@ -54,7 +54,7 @@ func HeaderURL() {
 		panic(err)
 	}
 
-	gouse.Println(gouse.MapAsString(header))
+	gouse.Println(gouse.Map2Str(header))
 }
 
 /*