+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1](https://github.com/thuongtruong109/gouse/releases/tag/v1.1) - 01/04/2025
+
+### What's new
+
+- Add packages (color, print, time...)
+- Update test cases, and benchmark tests
+- Fix compatible versions of dependencies
+
+[**Full Changelog**](https://github.com/thuongtruong109/gouse/commits/v1.1)
+
 ## [1.0](https://github.com/thuongtruong109/gouse/releases/tag/v1.0) - 18/12/2024
 
 ### What's new